+++ internal/services/parse/resource_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 )
 
-func Test_NewResourceID(t *testing.T) {
+func Test_ResourceIDWithResourceType(t *testing.T) {
 	testData := []struct {
 		ResourceId       string
 		ResourceType     string
@@ -36,6 +36,19 @@ func Test_NewResourceID(t *testing.T) {
 				AzureResourceId:   "/subscriptions/12345678-1234-9876-4563-123456789012/resourceGroups/test",
 				Name:              "test",
 				ParentId:          "/subscriptions/12345678-1234-9876-4563-123456789012",
+			},
+		},
+
+		{
+			ResourceType:     "Microsoft.Resources/subscriptions@2021-04-01",
+			ResourceId:       "/subscriptions/12345678-1234-9876-4563-123456789012",
+			ResourceDefExist: false,
+			Expected: &ResourceId{
+				ApiVersion:        "2021-04-01",
+				AzureResourceType: "Microsoft.Resources/subscriptions",
+				AzureResourceId:   "/subscriptions/12345678-1234-9876-4563-123456789012",
+				Name:              "12345678-1234-9876-4563-123456789012",
+				ParentId:          "/",
 			},
 		},
 
@@ -169,7 +182,7 @@ func Test_NewResourceID(t *testing.T) {
 	for _, v := range testData {
 		t.Logf("[DEBUG] Testing %q %q", v.ResourceId, v.ResourceType)
 
-		actual, err := NewResourceID(v.ResourceId, v.ResourceType)
+		actual, err := ResourceIDWithResourceType(v.ResourceId, v.ResourceType)
 		if err != nil {
 			if v.Error {
 				continue
@@ -199,13 +212,25 @@ func Test_NewResourceID(t *testing.T) {
 	}
 }
 
-func Test_ResourceID(t *testing.T) {
+func Test_ResourceIDWithApiVersion(t *testing.T) {
 	testData := []struct {
 		Input            string
 		Error            bool
 		ResourceDefExist bool
 		Expected         *ResourceId
 	}{
+		{
+			// tenant scope
+			Input:            "/subscriptions/12345678-1234-9876-4563-123456789012?api-version=2021-04-01",
+			ResourceDefExist: false,
+			Expected: &ResourceId{
+				ApiVersion:        "2021-04-01",
+				AzureResourceType: "Microsoft.Resources/subscriptions",
+				AzureResourceId:   "/subscriptions/12345678-1234-9876-4563-123456789012",
+				Name:              "12345678-1234-9876-4563-123456789012",
+				ParentId:          "/",
+			},
+		},
 
 		{
 			// tenant scope
@@ -354,7 +379,7 @@ func Test_ResourceID(t *testing.T) {
 	for _, v := range testData {
 		t.Logf("[DEBUG] Testing %q", v.Input)
 
-		actual, err := ResourceID(v.Input)
+		actual, err := ResourceIDWithApiVersion(v.Input)
 		if err != nil {
 			if v.Error {
 				continue
@@ -384,7 +409,7 @@ func Test_ResourceID(t *testing.T) {
 	}
 }
 
-func Test_BuildResourceID(t *testing.T) {
+func Test_NewResourceID(t *testing.T) {
 	testData := []struct {
 		Name             string
 		ParentId         string
@@ -393,6 +418,19 @@ func Test_BuildResourceID(t *testing.T) {
 		Error            bool
 		Expected         *ResourceId
 	}{
+		{
+			// tenant scope
+			Name:             "12345678-1234-9876-4563-123456789012",
+			ParentId:         "/",
+			ResourceType:     "Microsoft.Resources/subscriptions@2021-04-01",
+			ResourceDefExist: false,
+			Expected: &ResourceId{
+				ApiVersion:        "2021-04-01",
+				AzureResourceType: "Microsoft.Resources/subscriptions",
+				AzureResourceId:   "/subscriptions/12345678-1234-9876-4563-123456789012",
+			},
+		},
+
 		{
 			// tenant scope
 			Name:             "myDeployment",
@@ -593,7 +631,7 @@ func Test_BuildResourceID(t *testing.T) {
 	for _, v := range testData {
 		t.Logf("[DEBUG] Testing %q %q %q", v.Name, v.ParentId, v.ResourceType)
 
-		actual, err := BuildResourceID(v.Name, v.ParentId, v.ResourceType)
+		actual, err := NewResourceID(v.Name, v.ParentId, v.ResourceType)
 		if err != nil {
 			if v.Error {
 				continue